Extract Type enum to another file
